sample-apps/chat-model-server/app/requirements.txt
fastapi
uvicorn[standard]
datasets[s3]
transformers==4.18.0
torch-neuron
neuron-cc[tensorflow]
protobuf
torchvision
